reorder some render functions to be in the appropriate shape files
