+++ include/cylinder.h
@@ -5,8 +5,13 @@
 # include "minirt.h"
 # include "scene.h"
 # include "vector.h"
+
 # include <stdbool.h>
 
+void	compute_cylinder_light(t_vec3 *normal, t_point3 *intersect,
+	t_color *color, t_result *result);
+t_vec3	get_cylinder_normal(t_shape *cyl, t_point3 *intersect);
+void	handle_cylinder_intersect(double t[2], t_shape *cyl, t_range range, t_result *result);
 bool	hit_cylinder(t_point3 *origin, t_vec3 *dir, t_shape *cyl, double *t);
 void	create_cylinder(t_point3 pos, t_vec3 dir, float radius, float height, t_material mat);
