+++ include/sphere.h
@@ -1,6 +1,7 @@
 #ifndef SPHERE_H
 # define SPHERE_H
 
+# include "minirt.h"
 # include "scene.h"
 # include "point3.h"
 # include "material.h"
@@ -8,6 +9,8 @@
 # include <stdbool.h>
 
 // Functions //
+void	compute_sphere_light(t_vec3 *normal, t_point3 *intersect, t_color *color, t_result *result);
+void		handle_sphere_intersect(double t[2], t_shape *tmp, t_range range, t_result *result);
 int		create_sphere(t_point3 position, float radius, t_material mat);
 void	print_sphere(t_shape *sphere);
 bool	hit_sphere(t_point3 *origin, t_vec3 *dir, t_shape *sphere, double *t);
