+++ src/render.c
@@ -20,85 +20,6 @@
 // Reflection Depth, going over 3 barely adds anything visually
 #define MAXDEPTH 3
 
-/* TODO feel free to rearrange things, add static statements because the following functions were made
-	for norm and readability purposes. but they're just part of compute_light()
-*/
-
-t_vec3	get_cylinder_normal(t_shape *cyl, t_point3 *intersect)
-{
-	t_vec3		base_to_p;
-	t_vec3		proj;
-	t_vec3		normal;
-	double		axis_height;
-
-	base_to_p = point3_sub(intersect, &cyl->position);
-	axis_height = dot_product(&base_to_p, &cyl->direction);
-	proj = point3_scale(&cyl->direction, axis_height);
-	normal = point3_sub(&base_to_p, &proj);
-	vec_normalize(&normal);
-	return (normal);
-}
-
-void	compute_sphere_light(t_vec3 *normal, t_point3 *intersect, t_color *color, t_result *result)
-{
-	*normal = point3_sub(intersect, &result->closest->position);
-	vec_normalize(normal);
-	*color = result->closest->mat.color;
-}
-
-void	compute_plane_light(t_vec3 *normal, t_color *color, t_result *result)
-{
-	*normal = result->closest->normal;
-	*color = result->closest->mat.color;
-}
-
-void	compute_cylinder_light(t_vec3 *normal, t_point3 *intersect,
-								t_color *color, t_result *result)
-{
-	t_shape	*cyl;
-
-	cyl = result->closest;
-	*normal = get_cylinder_normal(cyl, intersect);
-	*color = cyl->mat.color;
-}
-
-void		handle_plane_intersect(double t[2], t_shape *tmp, t_range range, t_result *result)
-{
-	if (is_in_range(t[0], range) && t[0] < result->closest_t)
-	{
-		result->closest = tmp;
-		result->closest_t = t[0];
-	}
-}
-
-void		handle_sphere_intersect(double t[2], t_shape *tmp, t_range range, t_result *result)
-{
-	if (is_in_range(t[0], range) && t[0] < result->closest_t)
-	{
-		result->closest = tmp;
-		result->closest_t = t[0];
-	}
-	if (is_in_range(t[1], range) && t[1] < result->closest_t)
-	{
-		result->closest = tmp;
-		result->closest_t = t[1];
-	}
-}
-
-void	handle_cylinder_intersect(double t[2], t_shape *cyl, t_range range, t_result *result)
-{
-	if (is_in_range(t[0], range) && t[0] < result->closest_t)
-	{
-		result->closest = cyl;
-		result->closest_t = t[0];
-	}
-	if (is_in_range(t[1], range) && t[1] < result->closest_t)
-	{
-		result->closest = cyl;
-		result->closest_t = t[1];
-	}
-}
-
 /**
  * @brief Computes the closest intersection between a ray starting at `origin`
  * in direction `dir`, contained within `range`.
@@ -177,7 +98,7 @@ t_color	ray_color(t_point3 origin, t_vec3 dir, t_range t_range, int depth)
 	t_result	result;
 	t_color		local_color;
 	t_color		reflected;
-	
+
 	result = closest_intersect(&origin, &dir, t_range);
 	if (!result.closest)
 		return ((t_color)SKY_COLOR);
