+++ src/scene/plane.c
@@ -9,6 +9,21 @@
 #include <math.h>
 #include <assert.h>
 
+void	compute_plane_light(t_vec3 *normal, t_color *color, t_result *result)
+{
+	*normal = result->closest->normal;
+	*color = result->closest->mat.color;
+}
+
+void		handle_plane_intersect(double t[2], t_shape *tmp, t_range range, t_result *result)
+{
+	if (is_in_range(t[0], range) && t[0] < result->closest_t)
+	{
+		result->closest = tmp;
+		result->closest_t = t[0];
+	}
+}
+
 /**
  *	Find out if we are hitting a plane with our ray.
  */
@@ -16,7 +31,7 @@ bool hit_plane(t_point3 *origin, t_vec3 *dir, t_shape *plane, double *t)
 {
 	double	denom;
 	t_vec3	origin_to_plane;
-	
+
 	denom = dot_product(&plane->normal, dir);
 
 	if (fabs(denom) > 1e-6) // floatabsolute over 0.000001 or 10^-6
